+++ Cargo.toml
@@ -99,6 +99,7 @@ rolldown_plugin_wasm_helper             = { version = "0.1.0", path = "./crates/
 rolldown_resolver                       = { version = "0.1.0", path = "./crates/rolldown_resolver" }
 rolldown_rstr                           = { version = "0.1.0", path = "./crates/rolldown_rstr" }
 rolldown_sourcemap                      = { version = "0.1.0", path = "./crates/rolldown_sourcemap" }
+rolldown_std_utils                      = { version = "0.1.0", path = "./crates/rolldown_std_utils" }
 rolldown_testing                        = { version = "0.1.0", path = "./crates/rolldown_testing" }
 rolldown_testing_config                 = { version = "0.1.0", path = "./crates/rolldown_testing_config" }
 rolldown_tracing                        = { version = "0.1.0", path = "./crates/rolldown_tracing" }
