+++ crates/rolldown_utils/Cargo.toml
@@ -16,21 +16,22 @@ doctest = false
 workspace = true
 
 [dependencies]
-anyhow      = { workspace = true }
-base64-simd = { workspace = true }
-futures     = { workspace = true }
-glob-match  = { workspace = true }
-indexmap    = { workspace = true, features = ["serde"] }
-infer       = { workspace = true }
-memchr      = { workspace = true }
-mime        = { workspace = true }
-oxc         = { workspace = true }
-phf         = { workspace = true }
-regex       = { workspace = true }
-regress     = { workspace = true }
-rustc-hash  = { workspace = true }
-sugar_path  = { workspace = true }
-xxhash-rust = { workspace = true, features = ["xxh3"] }
+anyhow             = { workspace = true }
+base64-simd        = { workspace = true }
+futures            = { workspace = true }
+glob-match         = { workspace = true }
+indexmap           = { workspace = true, features = ["serde"] }
+infer              = { workspace = true }
+memchr             = { workspace = true }
+mime               = { workspace = true }
+oxc                = { workspace = true }
+phf                = { workspace = true }
+regex              = { workspace = true }
+regress            = { workspace = true }
+rolldown_std_utils = { workspace = true }
+rustc-hash         = { workspace = true }
+sugar_path         = { workspace = true }
+xxhash-rust        = { workspace = true, features = ["xxh3"] }
 
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
